backend/ats/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
pandas
scikit-learn
joblib
numpy
requests
PyPDF2
